pyproject.toml
[project]
name = "ebarimt-pos-sdk"
version = "0.1.4"
description = "Python SDK for Ebarimt POS API 3.0"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "SenergyTech" }]
dependencies = [
    "authlib>=1.6.8",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
]


[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "respx>=0.22.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
docs = ["sphinx>=8.1.3", "myst-parser>=4.0.1", "sphinx-rtd-theme>=3.1.0"]

[build-system]
requires = ["uv_build>=0.9.27,<0.10.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = ["integration: tests that require a real PosAPI server"]

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = ["tests"]
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]
